Improve network-script config parsing
The previous implementation of the network-scripts had a 'two pass'
model where recognized options were mapped to actual properties, and
everything else was dumped into the options hash. This doesn't handle
cases where a network-script option doesn't map neatly to a property.

This commit moves the parsing to a case statement that makes it easier
to handle this behavior.