Remove with_env, and trust in PATH being correct
Modern versions of Facter (v3) don't have .with_env().
In this first step of modernising our facts, we remove the .with_env()
block and simply trust that the PATH is correct and that we'll be able
to find ip(8).

In a next step, we should replace Facter::Util::Resolution with
Facter::Core::Execution. note that this will require reworking the spec
tests which rely on different sets of external facts that *do* use
Facter::Util::Resolution.