Fix issue with malformed interfaces file
When using inet and inet6, for example, the interfaces file was
considered malformed, even though both configurations can exist in
parallel.

This is non-total fix; one would technically need to check if there is
either a blacklist or whitelist of families that are/aren't compatible with
each other; but to my knowledge, inet and inet6 are the only significant
special cases in that regard.

And the whole construction of this provider is particularly difficult to
modify for the use case of multiple `Instance` objects of the same name…