Windows can't handle the : character in a filename.
This renames the file and then makes a symlink to the original filename
when the specs are run. This will allow the module to be installed
without errors onto a windows machine.

The other alternatives I considered were things like mocking File.read to
magically translate invalid characters, but that seemed like more wtf
than this. Happy to consider other approaches though!

Fixes #224