Allow for setting 'options' in network_route on RHEL-like OS's
Prior to this, the `network_route` type would not manage the `options`
property on RHEL-like OS's.

For example, on RHEL 7, the `options` property of `"table foo"` would be
completely ignored and not managed by Puppet:
```
network_route { '10.10.0.0/16':
  ensure    => 'present',
  gateway   => '10.10.0.5',
  interface => 'eth1',
  netmask   => '255.255.0.0',
  network   => '10.10.0.0',
  options   => 'table foo',
}
```

The problem was caused by a missing `feature` specification in the type.

This fixes that problem and allows `network_route`'s to use the
`options` property.