Add Debian 11 and 12 support
Change name of Debian package ifenslave-2.6 to ifenslave.

As of Debian 11, the package ifenslave-2.6 is no longer available,
causing puppet run failure when using bonds.

Debian 9 and 10 support is maintained because they also have the
ifenslave package (with ifenslave-2.6 marked as transitional).

This drops support for bonds in Debian 8, which is EOL since June 2018.